Create Comments
Comments
Comments don't do... anything. They are ignored by the Python interpreter. That said, they're good for what the name implies: adding comments to your code in plain English (or whatever language you speak).

Single Line Comment
A single # makes the rest of the line a comment:

# speed describes how fast the player
# moves in meters per second
speed = 2

Multi-Line Comments (Aka docstrings)
You can use triple quotes to start and end multi-line comments as well:

"""
    the code found below
    will print 'Hello, World!' to the console
"""
print("Hello, World!")

This is useful if you don't want to add the # to the start of each line when writing paragraphs of comments.

Assignment
Line #1 in the code was meant to be a comment, but the developer forgot to use the correct syntax (#).

Fix the bug by commenting out the first line of code.