Create Curse
Curse
An enemy's weapons can be cursed so that they don't deal as much damage.

Assignment
Complete the curse function. It accepts a weapon_damage parameter and returns two values:

The lesser_cursed damage: reduce the input weapon_damage from 100% to 50% (50% reduction).
The greater_cursed damage: reduce the input weapon_damage from 100% to 25% (75% reduction).
A greater curse is more powerful than a lesser curse, so it reduces the damage more.

Tip
You can multiply a number by a decimal to get a percentage of a number. For example:

30% of 50 is 50 * 0.3

Ignore the float function in the print statements. We will introduce floats later.